Initial changes to ssr

+++ kemenu-frontend/src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit, PLATFORM_ID } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { filter } from 'rxjs/operators';
 import { Router, NavigationEnd } from '@angular/router';
@@ -6,6 +6,7 @@ import { CookieService } from 'ngx-cookie-service';
 import { AuthenticationService } from '@services/authentication/authentication.service';
 import { Tokens } from '@models/auth/tokens.model';
 import Utils from './utils/utils';
+import { isPlatformBrowser } from '@angular/common';
 
 declare var gtag;
 
@@ -21,16 +22,21 @@ export class AppComponent implements OnInit {
   public cookieBASE64: string;
   public showCookiesAlert = true;
 
+  isBrowser: boolean;
+
   constructor(
     translate: TranslateService,
     private router: Router,
     private cookieService: CookieService,
-    private authService: AuthenticationService
+    private authService: AuthenticationService,
+    @Inject(PLATFORM_ID) platformId: Object
   ) {
-    translate.setDefaultLang(Utils.getBrowserLang());
-    translate.use(Utils.getBrowserLang());
+    this.isBrowser = isPlatformBrowser(platformId);
 
-    const navEndEvents$ = this.router.events
+    if(this.isBrowser) {
+      translate.setDefaultLang(Utils.getBrowserLang());
+      translate.use(Utils.getBrowserLang());
+      const navEndEvents$ = this.router.events
       .pipe(
         filter(event => event instanceof NavigationEnd)
       );
@@ -39,36 +45,44 @@ export class AppComponent implements OnInit {
         page_path: event.urlAfterRedirects
       });
     });
+    } 
+
+
   }
 
   ngOnInit() {
-    if (this.cookieService.get('show_menu')) {
-      localStorage.setItem('COOKIE-SHOW-MENU', this.cookieService.get('show_menu'));
-      this.router.navigateByUrl('/show');
+    if(this.isBrowser) {
+      if (this.cookieService.get('show_menu')) {
+        localStorage.setItem('COOKIE-SHOW-MENU', this.cookieService.get('show_menu'));
+        this.router.navigateByUrl('/show');
+      }
+  
+      if (this.cookieService.get('confirmed_email')) {
+        localStorage.setItem('COOKIE-CONFIRMED-EMAIL', this.cookieService.get('confirmed_email'));
+        this.cookieService.delete('confirmed_email');
+      }
+  
+      if (this.cookieService.get('forgot_password_email')) {
+        localStorage.setItem('FORGOT-PASSWORD-EMAIL', this.cookieService.get('forgot_password_email'));
+        this.cookieService.delete('forgot_password_email');
+        this.router.navigateByUrl('/changePassword');
+      }
+  
+      if (this.cookieService.get('CookieKemenu')) {
+        this.showCookiesAlert = false;
+      }
+  
+      if (localStorage.getItem(this.JWT_TOKEN)) {
+        this.checkExpirationToken();
+      }
     }
 
-    if (this.cookieService.get('confirmed_email')) {
-      localStorage.setItem('COOKIE-CONFIRMED-EMAIL', this.cookieService.get('confirmed_email'));
-      this.cookieService.delete('confirmed_email');
-    }
-
-    if (this.cookieService.get('forgot_password_email')) {
-      localStorage.setItem('FORGOT-PASSWORD-EMAIL', this.cookieService.get('forgot_password_email'));
-      this.cookieService.delete('forgot_password_email');
-      this.router.navigateByUrl('/changePassword');
-    }
-
-    if (this.cookieService.get('CookieKemenu')) {
-      this.showCookiesAlert = false;
-    }
-
-    if (localStorage.getItem(this.JWT_TOKEN)) {
-      this.checkExpirationToken();
-    }
   }
 
   onActivate() {
-    window.scroll(0, 0);
+    if(this.isBrowser) {
+      window.scroll(0, 0);
+    }
   }
 
   checkExpirationToken() {
